requirements.txt
django
gunicorn
django-heroku
flask
websockets
python-binance
datetime
config
ujson
numpy

